[ft-#157644385] Get all request for a user
